Refactor by splitting codebase formatting into separate folder
